# [Don't merge] export strict optimize for Windows.