--- a/torch/_export/__init__.py
+++ b/torch/_export/__init__.py
@@ -118,9 +118,6 @@ def capture_pre_autograd_graph(
 
     capture_pre_autograd_graph_warning()
 
-    if sys.platform == "win32":
-        raise RuntimeError("capture_pre_autograd_graph not yet supported on Windows")
-
     assert isinstance(f, torch.nn.Module), "Expected an nn.Module instance."
 
     if kwargs is None: